Additional exported RCausalML objects

Documentation placeholder for exported objects that are also documented in other function/class help pages.

Details

MultiArmCausalBoost is an R6 class for multi-arm causal boosting. multi_arm_ATE and multi_arm_PEHE are evaluation helpers used by that class. coef is the exported S3 generic with methods documented on their respective class help pages.

See also